Moves away from Combine for running one-time internal pipelines
