+++ Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/InternalPresentation.swift
@@ -16,7 +16,7 @@ public typealias PaywallStatePublisher = AnyPublisher<PaywallState, Never>
 typealias PresentablePipelineOutputPublisher = AnyPublisher<PresentablePipelineOutput, Error>
 
 extension Superwall {
-  /// Runs a combine pipeline to present a paywall, publishing ``PaywallState`` objects that provide updates on the lifecycle of the paywall.
+  /// Runs a background task to present a paywall, publishing ``PaywallState`` objects that provide updates on the lifecycle of the paywall.
   ///
   /// - Parameters:
   ///   - request: A presentation request of type `PresentationRequest` to feed into a presentation pipeline.
@@ -25,32 +25,35 @@ extension Superwall {
   @discardableResult
   func internallyPresent(
     _ request: PresentationRequest,
-    _ paywallStatePublisher: PassthroughSubject<PaywallState, Never> = .init()
+    _ publisher: PassthroughSubject<PaywallState, Never> = .init()
   ) -> PaywallStatePublisher {
-    let presentationSubject = PresentationSubject(request)
+    Task {
+      do {
+        try await checkNoPaywallAlreadyPresented(request, publisher)
+        await waitToPresent(request)
+        let debugInfo = logPresentation(request, "Called Superwall.shared.track")
+        try checkDebuggerPresentation(request, publisher)
+        let assignmentOutput = try await evaluateRules(request, debugInfo: debugInfo)
+        try await checkUserSubscription(request, assignmentOutput.triggerResult, publisher)
+        confirmHoldoutAssignment(input: assignmentOutput)
+        let triggerResultOutput = try await handleTriggerResult(request, assignmentOutput, publisher)
+        let paywallVcOutput = try await getPaywallViewController(request, triggerResultOutput, publisher)
 
-    presentationSubject
-      .eraseToAnyPublisher()
-      .checkNoPaywallAlreadyPresented(paywallStatePublisher)
-      .waitToPresent()
-      .logPresentation("Called Superwall.shared.track")
-      .checkDebuggerPresentation(paywallStatePublisher)
-      .evaluateRules()
-      .checkUserSubscription(paywallStatePublisher)
-      .confirmHoldoutAssignment()
-      .handleTriggerResult(paywallStatePublisher)
-      .getPaywallViewController(paywallStatePublisher)
-      .checkPaywallIsPresentable(paywallStatePublisher)
-      .confirmPaywallAssignment()
-      .presentPaywall(paywallStatePublisher)
-      .storePresentationObjects(presentationSubject, paywallStatePublisher)
-      .logErrors(from: request)
-      .subscribe(Subscribers.Sink(
-        receiveCompletion: { _ in },
-        receiveValue: { _ in }
-      ))
+        let presentableOutput = try await checkPaywallIsPresentable(
+          input: paywallVcOutput,
+          request: request,
+          publisher
+        )
+        
+        confirmPaywallAssignment(request: request, input: presentableOutput)
+        try await presentPaywall(request, presentableOutput, publisher)
+        storePresentationObjects(request, publisher)
+      } catch {
+        logErrors(from: request, error)
+      }
+    }
 
-    return paywallStatePublisher
+    return publisher
       .receive(on: DispatchQueue.main)
       .eraseToAnyPublisher()
   }
