--- Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/Operators/PresentPaywallOperator.swift
@@ -1,69 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Yusuf Tör on 26/09/2022.
-//
-
-import UIKit
-import Combine
-
-extension AnyPublisher where Output == PresentablePipelineOutput, Failure == Error {
-  /// Presents the paywall view controller, stores the presentation request for future use,
-  /// and sends back a `presented` state to the paywall state publisher.
-  ///
-  /// - Parameters:
-  ///   - paywallStatePublisher: A `PassthroughSubject` that gets sent ``PaywallState`` objects.
-  ///
-  /// - Returns: A publisher that contains info for the next pipeline operator.
-  func presentPaywall(
-    _ paywallStatePublisher: PassthroughSubject<PaywallState, Never>
-  ) -> PresentablePipelineOutputPublisher {
-    flatMap { input in
-      Future { promise in
-        Task {
-          Task.detached {
-            let trackedEvent = InternalSuperwallEvent.PresentationRequest(
-              eventData: input.request.presentationInfo.eventData,
-              type: input.request.flags.type,
-              status: .presentation,
-              statusReason: nil
-            )
-            await Superwall.shared.track(trackedEvent)
-          }
-          await MainActor.run {
-            input.paywallViewController.present(
-              on: input.presenter,
-              eventData: input.request.presentationInfo.eventData,
-              presentationStyleOverride: input.request.paywallOverrides?.presentationStyle,
-              paywallStatePublisher: paywallStatePublisher
-            ) { isPresented in
-              if isPresented {
-                let state: PaywallState = .presented(input.paywallViewController.paywallInfo)
-                paywallStatePublisher.send(state)
-                promise(.success(input))
-              } else {
-                Logger.debug(
-                  logLevel: .info,
-                  scope: .paywallPresentation,
-                  message: "Paywall Already Presented",
-                  info: input.debugInfo
-                )
-                let error = InternalPresentationLogic.presentationError(
-                  domain: "SWPresentationError",
-                  code: 102,
-                  title: "Paywall Already Presented",
-                  value: "Trying to present paywall while another paywall is presented."
-                )
-                paywallStatePublisher.send(.presentationError(error))
-                paywallStatePublisher.send(completion: .finished)
-                promise(.failure(PresentationPipelineError.paywallAlreadyPresented))
-              }
-            }
-          }
-        }
-      }
-    }
-    .eraseToAnyPublisher()
-  }
-}
