+++ Sources/SuperwallKit/Paywall/Presentation/Internal Presentation/Operators/StorePresentationObjects.swift
@@ -1,0 +1,29 @@
+//
+//  File.swift
+//  
+//
+//  Created by Yusuf Tör on 11/05/2023.
+//
+
+import Foundation
+import Combine
+
+extension Superwall {
+  /// Presents the paywall view controller, stores the presentation request for future use,
+  /// and sends back a `presented` state to the paywall state publisher.
+  ///
+  /// - Parameters:
+  ///   - paywallStatePublisher: A `PassthroughSubject` that gets sent ``PaywallState`` objects.
+  ///
+  /// - Returns: A publisher that contains info for the next pipeline operator.
+  func storePresentationObjects(
+    _ request: PresentationRequest,
+    _ paywallStatePublisher: PassthroughSubject<PaywallState, Never>
+  ) {
+    let lastPaywallPresentation = LastPresentationItems(
+      request: request,
+      statePublisher: paywallStatePublisher
+    )
+    presentationItems.last = lastPaywallPresentation
+  }
+}
